Abstracted MemShareController
